implement basic docker image building and tagging to ghcr.io
For forks that wish to self publish, or test as I needed to, they can
add ENABLE_PUSH=true in their GH repository variable configuration, and
it'll enable pushing to their ghcr.io repository.

For everyone else, this just kicks the docker build of the image to
ensure it's buildable.

As to the image: this uses a minimalist but fast rust static server.
This keeps the image size minimal.  This is also deprived to a UID of 3000,
which however does require the user to align user IDs.

This is preferable to running the processes as root, but admittedly isn't
optimal for user experience.

This also does a test check before doing the push, for basic sanity.

Signed-off-by: Brian Harring <ferringb@gmail.com>